+++ lib/glimmer-dsl-web/samples/hello/hello_style.rb
@@ -22,6 +22,10 @@
 require 'glimmer-dsl-web'
 
 class ButtonModel
+  BUTTON_STYLE_ATTRIBUTES = [
+    :width, :height, :font_size,
+    :background_color_red, :background_color_green, :background_color_blue
+  ]
   WIDTH_MIN = 160
   WIDTH_MAX = 960
   HEIGHT_MIN = 100
@@ -29,13 +33,16 @@ class ButtonModel
   FONT_SIZE_MIN = 40
   FONT_SIZE_MAX = 200
   
-  attr_accessor :text, :pushed, :width, :height, :font_size
+  attr_accessor :text, :pushed, *BUTTON_STYLE_ATTRIBUTES
   
   def initialize
     @text = 'Push'
     @width = WIDTH_MIN
     @height = HEIGHT_MIN
     @font_size = FONT_SIZE_MIN
+    @background_color_red = 173
+    @background_color_green = 216
+    @background_color_blue = 230
   end
   
   def push
@@ -76,17 +83,11 @@ class StyledButton
                       on_read: ->(pushed) { pushed ? 'pushed' : 'pulled' }
                     ]
       
-      style <= [ button_model, :width,
-                 on_read: method(:button_style_value) # convert value on read before storing in style
-               ]
-      
-      style <= [ button_model, :height,
-                 on_read: method(:button_style_value) # convert value on read before storing in style
-               ]
-      
-      style <= [ button_model, :font_size,
-                 on_read: method(:button_style_value) # convert value on read before storing in style
-               ]
+      ButtonModel::BUTTON_STYLE_ATTRIBUTES.each do |attribute|
+        style <= [ button_model, attribute,
+                   on_read: method(:button_style_value) # convert value on read before storing in style
+                 ]
+      end
       
       onclick do
         button_model.push
@@ -94,30 +95,30 @@ class StyledButton
     }
   }
   
-  style {'
-    button {
+  style {"
+    .#{component_element_class} {
       font-family: Courrier New, Courrier;
       border-radius: 5px;
       border-width: 17px;
-      border-color: #ACC7D5;
-      background-color: #ADD8E6;
       margin: 5px;
     }
     
-    button.pulled {
+    .#{component_element_class}.pulled {
       border-style: outset;
     }
     
-    button.pushed {
+    .#{component_element_class}.pushed {
       border-style: inset;
     }
-  '}
+  "}
   
   def button_style_value
     "
       width: #{button_model.width}px;
       height: #{button_model.height}px;
       font-size: #{button_model.font_size}px;
+      background-color: rgb(#{button_model.background_color_red}, #{button_model.background_color_green}, #{button_model.background_color_blue});
+      border-color: rgb(#{button_model.background_color_red - 10}, #{button_model.background_color_green - 10}, #{button_model.background_color_blue - 10});
     "
   end
 end
@@ -137,6 +138,22 @@ class StyledButtonRangeInput
   }
 end
 
+class StyledButtonColorRangeInput
+  include Glimmer::Web::Component
+  
+  COLORS = [:red, :green, :blue]
+  
+  option :button_model
+  option :property
+  option :color
+  
+  markup {
+    input(type: 'range', min: 0, max: 255) {
+      value <=> [button_model, "#{property}_#{color}"]
+    }
+  }
+end
+
 class HelloStyle
   include Glimmer::Web::Component
   
@@ -145,32 +162,54 @@ class HelloStyle
   end
   
   markup {
-    div(class: 'hello-style') {
-      div(class: 'form-row') {
-        label('Styled Button Width:', for: 'styled-button-width-input')
+    div {
+      div(class: 'styled-button-form') {
+        label('Styled Button Width:', class: 'property-label', for: 'styled-button-width-input')
         styled_button_range_input(button_model: @button_model, property: :width, property_min: ButtonModel::WIDTH_MIN, property_max: ButtonModel::WIDTH_MAX, id: 'styled-button-width-input')
-      }
-      div(class: 'form-row') {
-        label('Styled Button Height:', for: 'styled-button-height-input')
+        
+        label('Styled Button Height:', class: 'property-label', for: 'styled-button-height-input')
         styled_button_range_input(button_model: @button_model, property: :height, property_min: ButtonModel::HEIGHT_MIN, property_max: ButtonModel::HEIGHT_MAX, id: 'styled-button-height-input')
-      }
-      div(class: 'form-row') {
-        label('Styled Button Font Size:', for: 'styled-button-font-size-input')
+        
+        label('Styled Button Font Size:', class: 'property-label', for: 'styled-button-font-size-input')
         styled_button_range_input(button_model: @button_model, property: :font_size, property_min: ButtonModel::FONT_SIZE_MIN, property_max: ButtonModel::FONT_SIZE_MAX, id: 'styled-button-font-size-input')
+        
+        label('Styled Button Background Color:', for: 'styled-button-background-color-input')
+        
+        StyledButtonColorRangeInput::COLORS.each do |color|
+          input_id = "styled-button-background-color-#{color}-input"
+          label("#{color.to_s.capitalize}:", class: 'color-label', for: input_id)
+          styled_button_color_range_input(button_model: @button_model, property: :background_color, color: color, id: input_id)
+        end
       }
+      
       styled_button(button_model: @button_model)
     }
   }
   
-  style {'
-    .hello-style {
+  style {"
+    .styled-button-form {
       padding: 20px;
+      display: inline-grid;
+      grid-template-columns: auto auto;
     }
     
-    .hello-style .form-row {
-      margin: 10px 0;
+    .styled-button-form label, input {
+      display: block;
+      margin: 5px 5px 5px 0;
     }
-  '}
+    
+    .styled-button-form label[for=styled-button-background-color-input] {
+      grid-column: 1 / 3;
+    }
+    
+    .styled-button-form .color-label {
+      margin-left: 5px;
+    }
+    
+    .#{component_element_class} .styled-button {
+      display: block;
+    }
+  "}
 end
 
 Document.ready? do